Move 1d rservoir code into subdir
